Add examples of Select using a slice of q.Matcher (#1)
I wanted to add some examples showing how conditionally someone could append various q.Matchers to a filter slice and showing you that err is thrown when no data is found on a Find.  This is a helpful example to see how to compose a query where logic in your algorithm will append filters based on various conditions since technically it seems like you can only call Select once to build the And's and cannot modify it after.  The slice example is how I am going to need implement things because not all my callers will know 100% the full filter when it's executed so it needs to be built up over time and then the slice of q.Matcher is passed to Select().

In this case of the data2 block, if I merely re-used data, the records would still exist because when storm cant find anything, it keeps your referenced data in there and doesnt clear it.  So anyone using this, make sure you handle errors like I did with the "not found" check so that you typically wont want to return error back on a not found when filtering data, rather just pass back the empty slice of data back to your caller who is making the query.